+++ frontend/src/apis/meetings/confirms.ts
@@ -1,6 +1,6 @@
 import { BASE_URL } from '@constants/api';
 
-import { fetchClient } from './_common/fetchClient';
+import { fetchClient } from '../_common/fetchClient';
 import type { MeetingType } from './meetings';
 
 export interface ConfirmDates {
