+++ frontend/src/hooks/useMeetingType/useMeetingType.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 
-import type { MeetingType } from '@apis/meetings';
+import type { MeetingType } from '@apis/meetings/meetings';
 
 export default function useMeetingType() {
   const [isChecked, setIsChecked] = useState(false);
